Refactoring MediaVersionUpload
