+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaVersionUpload/CropOverlay.js
