+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaVersionUpload/FocusPointOverlay.js
