+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaVersionUpload/README.md
@@ -1,0 +1,6 @@
+This component shows a thumbnail of the recent version of the uploaded media.
+Besides that, two buttons are rendered below which allow to open either the `FocusPointOverlay` or the `CropOverlay`. 
+The `FocusPointOverlay` makes it possible to set a focus point within the uploaded media element and the 
+`CropOverlay` provides a typical crop functionality.
+
+A `ResourceStore` component has to be passed so the component knows which locale, id and so on to use.
