+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaVersionUpload/cropOverlay.scss
