+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaVersionUpload/focusPointOverlay.scss
