+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaVersionUpload/index.js
@@ -1,0 +1,4 @@
+// @flow
+import MediaVersionUpload from './MediaVersionUpload';
+
+export default MediaVersionUpload;
