+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaVersionUpload/mediaVersionUpload.scss
@@ -1,0 +1,9 @@
+.buttons {
+    display: flex;
+    flex-wrap: wrap;
+
+    > * {
+        flex: 1;
+        margin-top: 20px;
+    }
+}
