--- src/Sulu/Bundle/MediaBundle/Resources/js/views/MediaDetails/README.md
@@ -1,7 +1,0 @@
-The `MediaDetails` view is registered with the key `sulu_media.details`. It shows a thumbnail of the recent version of the
-uploaded media. Besides that a form is rendered to edit information about the media.
-
-| Option        | Description                                                                                         |
-|---------------|-----------------------------------------------------------------------------------------------------|
-| locales       | Defines which locales are available in the locale chooser of this form.                             |
-| resourceStore | The store which allows to save and load data from the given type of resources                       |
