--- src/Sulu/Bundle/MediaBundle/Resources/js/views/MediaDetails/index.js
@@ -1,4 +1,0 @@
-// @flow
-import MediaDetails from './MediaDetails';
-
-export default MediaDetails;
