--- src/Sulu/Bundle/MediaBundle/Resources/js/views/MediaDetails/mediaDetails.scss
@@ -1,24 +1,0 @@
-@import 'sulu-admin-bundle/containers/Application/variables.scss';
-
-.media-detail {
-    max-width: $viewMaxWidth;
-    min-width: $viewMinWidth;
-}
-
-.form {
-    margin-bottom: 30px;
-}
-
-.image-section {
-    margin-bottom: 30px;
-}
-
-.buttons {
-    display: flex;
-    flex-wrap: wrap;
-
-    > * {
-        flex: 1;
-        margin-top: 20px;
-    }
-}
